pyproject.toml
[tool.poetry]
name = "court_parser"
version = "0.1.0"
description = ""
authors = ["la-strole <la-strole.sovaleto48@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
selenium = "^4.2.0"
lxml = "^4.9.0"
schedule = "^1.1.0"
pyTelegramBotAPI = "^4.5.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
